--- src/load/requirements.txt
@@ -1,27 +1,0 @@
-astroid==3.3.11
-boto3==1.39.10
-botocore==1.39.10
-dill==0.4.0
-dotenv==0.9.9
-iniconfig==2.1.0
-isort==6.0.1
-jmespath==1.0.1
-mccabe==0.7.0
-numpy==2.3.1
-packaging==25.0
-pandas==2.3.1
-platformdirs==4.3.8
-pluggy==1.6.0
-Pygments==2.19.2
-pylint==3.3.7
-pymssql==2.3.7
-pyodbc==5.2.0
-pytest==8.4.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.1
-pytz==2025.2
-s3transfer==0.13.1
-six==1.17.0
-tomlkit==0.13.3
-tzdata==2025.2
-urllib3==2.5.0
